donner la Permission seulement aux roles admin et auteur pour l'ensemble des foncationnalité du blog notament pour les articles, categories et tags , et restreindre le role auteur à la fonctionnalité Supprimer et le role user peut seulement créer son compte et regarder les articles

+++ resources/views/categories/index.blade.php
@@ -24,11 +24,13 @@
                         <th>{{ $categorie->id }}</th>
                         <td> <a href="{{route('categories.edit', $categorie->id)}}">{{ $categorie->name }}</a> </td>        
                         <td>    
+                            @can('delete-users')
+                              
                                 {!! Form::open(['route'=>['categories.destroy',$categorie->id],'method'=>'DELETE', ]) !!}
                                 {!! Form::submit('Delete', ['class'=>'btn btn-danger btn-sm  float-left']) !!}
             
                                 {!! Form::close() !!}
-                            
+                            @endcan
                         </td>
                     </tr>
                     @endforeach
